chore: refactor SettingsStore internals to handle breaking out the SettingsDialogActivity
