+++ src/main/AndroidManifest.xml
@@ -49,7 +49,7 @@
         android:launchMode="singleTop"/>
     <activity android:name="FreeSpaceWarningActivity"
         android:screenOrientation="portrait"/>
-    <activity android:name="SettingsDialogActivity"
+    <activity android:name=".components.settings_dialog.SettingsDialogActivity"
         android:screenOrientation="portrait"/>
     <activity android:name="RequestLocationPermissionActivity"
         android:screenOrientation="portrait"/>
